[REP] Update postCreateCommand in devcontainer.json
This commit updates the `postCreateCommand` in the `.devcontainer/devcontainer.json` file. The change is specifically on line 8.

Previously, the command was installing the `nm-lib` package from the `nm_lib` repository:
- "postCreateCommand": "python3 -m pip install -e git+https://github.com/AST-Course/nm_lib.git#egg=nm-lib"

Now, it installs the `nm-lib` package from the `AST5110` repository:
+ "postCreateCommand": "python3 -m pip install -e git+https://github.com/AST-Course/AST5110.git#egg=nm-lib"

This change ensures that the correct repository is used when setting up the development container.